updated leadModel seeder
